Testing version 0.72 - removing some small bugs. wxDataFile contains numer of samples. This feature currently is used nowhere.

+++ inc/wxCueFile/wxIndex.h
@@ -12,13 +12,13 @@ class wxIndex :public wxObject
 
 protected:
 
-	unsigned int m_number;
+	size_t m_number;
 	wxULongLong m_offset;
 	bool m_bCdFrames;
 
 protected:
 
-	void copy(const wxIndex&);
+	void copy( const wxIndex& );
 	const wxULongLong& GetOffset() const;
 	wxIndex& operator-=( wxULongLong );
 	wxIndex& operator+=( wxULongLong );
@@ -27,13 +27,13 @@ public:
 
 	static int CompareFn( wxIndex**, wxIndex** );
 
-	unsigned int GetNumber() const;
+	size_t GetNumber() const;
 	bool HasCdFrames() const;
 
 	wxIndex& SetNumber( unsigned int );
 	wxIndex& SetOffset( wxULongLong );
-	wxIndex& Assign( unsigned int, wxULongLong );
-	wxIndex& Assign( unsigned int, unsigned long, unsigned long, unsigned long );
+	wxIndex& Assign( size_t, wxULongLong );
+	wxIndex& Assign( size_t, unsigned long, unsigned long, unsigned long );
 
 	bool IsValid(bool = false) const;
 	static void FixDecimalPoint( wxString& );
@@ -43,7 +43,7 @@ public:
 	wxIndex(void);
 	wxIndex( unsigned int, wxULongLong );
 	wxIndex( const wxIndex& );
-	wxIndex& operator=(const wxIndex&);
+	wxIndex& operator=( const wxIndex& );
 
 	static bool TimeSpanToMsf( wxTimeSpan, unsigned int&, unsigned int&, unsigned int& );
 	static wxString GetTimeStr( unsigned int, unsigned int, double );
